Fix digest handling after 5ae680e49
This adds a sha256_fd() and fd support to rsa_verify(). This allows
passing the temporary fd directly to these functions to check
the signature. sha256_file() is now a wrapper around sha256_fd().

+++ libpkg/update.c
@@ -122,7 +122,7 @@ repo_archive_extract_file(int fd, const char *file, const char *dest, const char
 
 	while (archive_read_next_header(a, &ae) == ARCHIVE_OK) {
 		if (strcmp(archive_entry_pathname(ae), file) == 0) {
-			if (dest != NULL) {
+			if (dest_fd == -1) {
 				archive_entry_set_pathname(ae, dest);
 				/*
 				 * The repo should be owned by root and not writable
@@ -136,8 +136,7 @@ repo_archive_extract_file(int fd, const char *file, const char *dest, const char
 					rc = EPKG_FATAL;
 					goto cleanup;
 				}
-			}
-			if (dest_fd != -1) {
+			} else {
 				if (archive_read_data_into_fd(a, dest_fd) != 0) {
 					pkg_emit_errno("archive_read_extract", "extract error");
 					rc = EPKG_FATAL;
@@ -156,11 +155,12 @@ repo_archive_extract_file(int fd, const char *file, const char *dest, const char
 	if (repokey != NULL) {
 		if (sig != NULL) {
 			ret = rsa_verify(dest, repokey,
-					sig, siglen - 1);
+					sig, siglen - 1, dest_fd);
 			if (ret != EPKG_OK) {
 				pkg_emit_error("Invalid signature, "
 						"removing repository.");
-				unlink(dest);
+				if (dest != NULL)
+					unlink(dest);
 				free(sig);
 				rc = EPKG_FATAL;
 				goto cleanup;
@@ -170,7 +170,8 @@ repo_archive_extract_file(int fd, const char *file, const char *dest, const char
 			pkg_emit_error("No signature found in the repository.  "
 					"Can not validate against %s key.", repokey);
 			rc = EPKG_FATAL;
-			unlink(dest);
+			if (dest != NULL)
+				unlink(dest);
 			goto cleanup;
 		}
 	}
@@ -190,7 +191,6 @@ repo_fetch_remote_extract_tmp(struct pkg_repo *repo, const char *filename,
 	FILE *res = NULL;
 	const char *tmpdir;
 	char tmp[MAXPATHLEN];
-	char tmp_rsa[MAXPATHLEN];
 
 	fd = repo_fetch_remote_tmp(repo, filename, extension, t, rc);
 	if (fd == -1) {
@@ -205,7 +205,6 @@ repo_fetch_remote_extract_tmp(struct pkg_repo *repo, const char *filename,
 	mask = umask(022);
 	dest_fd = mkstemp(tmp);
 	umask(mask);
-	snprintf(tmp_rsa, MAXPATHLEN, "%s.rsa", tmp);
 	if (dest_fd == -1) {
 		pkg_emit_error("Could not create temporary file %s, "
 				"aborting update.\n", tmp);
@@ -213,12 +212,10 @@ repo_fetch_remote_extract_tmp(struct pkg_repo *repo, const char *filename,
 		goto cleanup;
 	}
 	(void)unlink(tmp);
-	if (repo_archive_extract_file(fd, archive_file, tmp_rsa, repo->pubkey, dest_fd) != EPKG_OK) {
+	if (repo_archive_extract_file(fd, archive_file, NULL, repo->pubkey, dest_fd) != EPKG_OK) {
 		*rc = EPKG_FATAL;
-		(void)unlink(tmp_rsa);
 		goto cleanup;
 	}
-	(void)unlink(tmp_rsa);
 
 	res = fdopen(dest_fd, "r");
 	if (res == NULL) {
